+++ crates/derive/src/batch/span_batch/tx_data/eip2930.rs
@@ -37,7 +37,7 @@ impl SpanBatchEip2930TransactionData {
                     SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
                 })?,
             ),
-            gas_limit: gas as u128,
+            gas_limit: gas,
             to: to.map_or(TxKind::Create, TxKind::Call),
             value: self.value,
             input: self.data.clone().into(),
