+++ crates/derive/src/batch/span_batch/tx_data/legacy.rs
@@ -34,7 +34,7 @@ impl SpanBatchLegacyTransactionData {
                     SpanBatchError::Decoding(SpanDecodingError::InvalidTransactionData)
                 })?,
             ),
-            gas_limit: gas as u128,
+            gas_limit: gas,
             to: to.map_or(TxKind::Create, TxKind::Call),
             value: self.value,
             input: self.data.clone().into(),
